Moved representations for journal in separate file
